requirements.txt
Flask==1.1.2
numpy==1.21.2
pandas==1.3.3
Pillow==9.0.0
Werkzeug==1.0.1
gunicorn==20.1.0
Jinja2==3.0.1
